cleanup from comments and remove temporary data file
